04-processing-snowpark/requirements.txt
snowflake-snowpark-python==1.16.0
snowflake-connector-python==3.10.1

